refactor(bootstrap): move app-wide HTTP middleware
